+++ packages/job-worker/src/ingest/generationSegment.ts
@@ -18,7 +18,7 @@ import { getSegmentId, getPartId, getRundown, canSegmentBeUpdated } from './lib'
 import { JobContext } from '../jobs'
 import { CommitIngestData } from './lock'
 import { BlueprintResultSegment, NoteSeverity } from '@sofie-automation/blueprints-integration'
-import { calculatePartExpectedDurationWithPreroll } from '@sofie-automation/corelib/dist/playout/timings'
+import { calculatePartExpectedDurationWithTransition } from '@sofie-automation/corelib/dist/playout/timings'
 import { wrapTranslatableMessageFromBlueprints } from '@sofie-automation/corelib/dist/TranslatableMessage'
 import { ReadOnlyCache } from '../cache/CacheBase'
 
@@ -216,7 +216,7 @@ export async function calculateSegmentsFromIngestData(
 						  }
 						: undefined,
 
-					expectedDurationWithPreroll: undefined, // Below
+					expectedDurationWithTransition: undefined, // Below
 				})
 				res.parts.push(part)
 
@@ -250,7 +250,7 @@ export async function calculateSegmentsFromIngestData(
 					)
 				)
 
-				part.expectedDurationWithPreroll = calculatePartExpectedDurationWithPreroll(part, processedPieces)
+				part.expectedDurationWithTransition = calculatePartExpectedDurationWithTransition(part, processedPieces)
 			})
 
 			preserveOrphanedSegmentPositionInRundown(context, cache, newSegment)