+++ packages/job-worker/src/playout/lookahead/__tests__/util.test.ts
@@ -115,7 +115,7 @@ describe('getOrderedPartsAfterPlayhead', () => {
 				_rank: index,
 				externalId: 'MOCK_PART_' + index,
 				title: 'Part ' + index,
-				expectedDurationWithPreroll: undefined,
+				expectedDurationWithTransition: undefined,
 			}
 		}